edgeai-modelmaker/requirements_freeze.txt
absl-py==2.3.1
addict==2.4.0
aenum==3.1.16
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chumpy==0.70
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
colored==2.3.1
coloredlogs==15.0.1
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.1
cycler==0.12.1
Cython==3.1.4
debugpy==1.8.17
decorator==5.2.1
descartes==1.1.0
dill==0.4.0
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.14.0
fire==0.7.1
flatbuffers==1.12
fonttools==4.60.0
fsspec==2024.6.1
google-auth==2.40.3
google-auth-oauthlib==0.4.6
graphviz==0.21
grpcio==1.75.0
h5py==3.14.0
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.5.2
json-tricks==3.17.3
kiwisolver==1.4.9
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mmcv==2.2.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
multiprocess==0.70.18
munkres==1.1.4
networkx==3.3
ninja==1.13.0
numpy==1.23.0
nuscenes-devkit==1.2.0
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
oauthlib==3.3.1
onnx==1.14.0
onnx-graphsurgeon==0.3.26
#onnxscript==0.2.5
onnxsim==0.4.35
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.2
parameterized==0.9.0
pillow==11.0.0
Pillow-SIMD==9.5.0.post2
platformdirs==4.4.0
pluggy==1.6.0
plyfile==1.1.2
prettytable==3.16.0
progiter==2.0.0
progressbar==2.5
protobuf==3.20.3
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.23
pycryptodome==3.23.0
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.10.0
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==3.1.0
terminaltables==3.1.10
tflite==2.18.0
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.4.0+cu124
torcheval==0.0.7
torchinfo==1.8.0
torchvision==0.19.0+cu124
tornado==6.5.2
tqdm==4.65.2
transforms3d==0.4.2
triton==3.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.14
Werkzeug==3.1.3
wurlitzer==3.1.1
xtcocotools==1.14.3
yapf==0.43.0
